pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "adkx"
version = "0.1.0"
description = "ADK extended Server"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Koichi Ozaki" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "google-adk>=1.27.1",
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/Koichi73/adkx"
Repository = "https://github.com/Koichi73/adkx"
Issues = "https://github.com/Koichi73/adkx/issues"

[dependency-groups]
dev = [
    "httpx>=0.28",
    "pytest>=9.0",
    "pytest-asyncio>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/adkx"]

[tool.hatch.build.targets.sdist]
include = ["src/adkx"]
